+++ src/Utility/TextFormatUtility.php
@@ -21,7 +21,7 @@ class TextFormatUtility
 {
     public static function formatNumber($unitPrice, $apiRoundingPrecision, $docPoint = '.', $thousandSep = '')
     {
-        return number_format($unitPrice, $apiRoundingPrecision, $docPoint, $thousandSep);
+        return number_format((float) $unitPrice, $apiRoundingPrecision, $docPoint, $thousandSep);
     }
 
     /**